Constant name compatibility file.
I look for the nonexistence of a pixel format definition that only
occurs in newer versions to trigger the inclusion of a file that defines
new names as the older versions.

This check is fragile, but the build with an old library probably won't
work anyway due to FFI imports failing to link. However, these name
definitions at least allow the Haskell library to compile, which is a
useful sanity check with something like Travis-CI.

Note that a warning is printed when we expect the build to be useless.