Removing nameCompat.h
This let us build ffmpeg-light on Travis CI, but it was a gross hack to
just let the compiler work by providing dummy definitions of things in
ffmpeg added more recently than the ancient version available on the
Travis infrastructure.

Maintaining these shims so as to work with multiple versions of ffmpeg
is not worth the trouble.