Fix support with older versions of FFmpeg
The previous approach didn't percolate the CPP macro from the .hsc
code to a downstream .hs module. We need to guard some parts of the
`Encode` module based on the version of FFmpeg, so we will for now
run `Encode` through `hsc2hs` so that it can directly `#include` a
header with the relevant version information.