Fix webcam support on Linux
This requires that we set a format string to "v4l2" when opening the
camera. Complicating matters is the usage of the "mjpeg" format string
that can offer higher frame rates and resolutions on common
webcams. Usage is simple: specify "mjpeg" as the format
string. Internally, we explicitly attach an MJPEG decoder to a v4l2
input format.

The demo program's "cam" variant now takes an optional second argument
to identify which camera device to use. On Linux, the default is
`/dev/video0`, but this can be overridden by running the program as,
for example, `cabal run demo -- cam 1` to use the device at
`/dev/video1` for recording.

Fixes #52