Fix webcam support on Linux
This requires that we set a format string to "v4l2" when opening the
camera. Complicating matters is the usage of the "mjpeg" format string
that can offer higher frame rates and resolutions on common
webcams. Usage is simple: specify "mjpeg" as the format
string. Internally, we explicitly attach an MJPEG decoder to a v4l2
input format.

Fixes #52