bump to sdl2-2.1.3.1
by bumping to a newer (but still very old at this point) LTS, we bumped
all our dependencies except those which are pinned in the extra-deps
section. There was an incompatibility between one of the dependencies in
that extra-deps section, sdl2-2.1.3, and one of the dependencies in the
LTS, vector-0.12, so I bumped to sdl2-2.1.3.1, which has a looser upper
bound for vector.